@@ -42,7 +42,7 @@ class _MyAppState extends State<MyApp> {
                   child: SizedBox.expand(
                     child: TextButton(
                       style: ButtonStyle(
-                        backgroundColor: MaterialStateProperty.all(Colors.blue),
+                        backgroundColor: WidgetStateProperty.all(Colors.blue),
                       ),
                       onPressed: _takePhoto,
                       child: Text(firstButtonText,
@@ -58,7 +58,7 @@ class _MyAppState extends State<MyApp> {
                     child: SizedBox.expand(
                   child: TextButton(
                     style: ButtonStyle(
-                      backgroundColor: MaterialStateProperty.all(Colors.white),
+                      backgroundColor: WidgetStateProperty.all(Colors.white),
                     ),
                     onPressed: _recordVideo,
                     child: Text(secondButtonText,
@@ -77,14 +77,14 @@ class _MyAppState extends State<MyApp> {
 
   void _takePhoto() async {
     ImagePicker()
-        .getImage(source: ImageSource.camera)
-        .then((PickedFile recordedImage) {
+        .pickImage(source: ImageSource.camera)
+        .then((XFile? recordedImage) {
       if (recordedImage != null && recordedImage.path != null) {
         setState(() {
           firstButtonText = 'saving in progress...';
         });
         GallerySaver.saveImage(recordedImage.path, albumName: albumName)
-            .then((bool success) {
+            .then((bool? success) {
           setState(() {
             firstButtonText = 'image saved!';
           });
@@ -95,14 +95,14 @@ class _MyAppState extends State<MyApp> {
 
   void _recordVideo() async {
     ImagePicker()
-        .getVideo(source: ImageSource.camera)
-        .then((PickedFile recordedVideo) {
+        .pickVideo(source: ImageSource.camera)
+        .then((XFile? recordedVideo) {
       if (recordedVideo != null && recordedVideo.path != null) {
         setState(() {
           secondButtonText = 'saving in progress...';
         });
         GallerySaver.saveVideo(recordedVideo.path, albumName: albumName)
-            .then((bool success) {
+            .then((bool? success) {
           setState(() {
             secondButtonText = 'video saved!';
           });
@@ -115,7 +115,7 @@ class _MyAppState extends State<MyApp> {
   void _saveNetworkVideo() async {
     String path =
         'https://sample-videos.com/video123/mp4/720/big_buck_bunny_720p_1mb.mp4';
-    GallerySaver.saveVideo(path, albumName: albumName).then((bool success) {
+    GallerySaver.saveVideo(path, albumName: albumName).then((bool? success) {
       setState(() {
         print('Video is saved');
       });
@@ -126,7 +126,7 @@ class _MyAppState extends State<MyApp> {
   void _saveNetworkImage() async {
     String path =
         'https://image.shutterstock.com/image-photo/montreal-canada-july-11-2019-600w-1450023539.jpg';
-    GallerySaver.saveImage(path, albumName: albumName).then((bool success) {
+    GallerySaver.saveImage(path, albumName: albumName).then((bool? success) {
       setState(() {
         print('Image is saved');
       });
@@ -153,7 +153,7 @@ class _ScreenshotWidgetState extends State<ScreenshotWidget> {
           child: SizedBox.expand(
             child: TextButton(
               style: ButtonStyle(
-                backgroundColor: MaterialStateProperty.all(Colors.pink),
+                backgroundColor: WidgetStateProperty.all(Colors.pink),
               ),
               onPressed: _saveScreenshot,
               child: Text(screenshotButtonText,
@@ -171,25 +171,30 @@ class _ScreenshotWidgetState extends State<ScreenshotWidget> {
     });
     try {
       //extract bytes
-      final RenderRepaintBoundary boundary =
-          _globalKey.currentContext.findRenderObject();
-      final ui.Image image = await boundary.toImage(pixelRatio: 3.0);
-      final ByteData byteData =
-          await image.toByteData(format: ui.ImageByteFormat.png);
-      final Uint8List pngBytes = byteData.buffer.asUint8List();
+      final RenderRepaintBoundary? boundary = _globalKey.currentContext
+          ?.findRenderObject() as RenderRepaintBoundary?;
+      if (boundary != null) {
+        final ui.Image image = await boundary.toImage(pixelRatio: 3.0);
 
-      //create file
-      final String dir = (await getApplicationDocumentsDirectory()).path;
-      final String fullPath = '$dir/${DateTime.now().millisecond}.png';
-      File capturedFile = File(fullPath);
-      await capturedFile.writeAsBytes(pngBytes);
-      print(capturedFile.path);
+        final ByteData? byteData =
+            await image.toByteData(format: ui.ImageByteFormat.png);
+        final Uint8List pngBytes = byteData!.buffer.asUint8List();
 
-      await GallerySaver.saveImage(capturedFile.path).then((value) {
-        setState(() {
-          screenshotButtonText = 'screenshot saved!';
+        //create file
+        final String dir = (await getApplicationDocumentsDirectory()).path;
+        final String fullPath = '$dir/${DateTime.now().millisecond}.png';
+        File capturedFile = File(fullPath);
+        await capturedFile.writeAsBytes(pngBytes);
+        print(capturedFile.path);
+
+        await GallerySaver.saveImage(capturedFile.path).then((value) {
+          setState(() {
+            screenshotButtonText = 'screenshot saved!';
+          });
         });
-      });
+      } else {
+        print("Render object is not a RenderRepaintBoundary or is null");
+      }
     } catch (e) {
       print(e);
     }
